работа над run check
