+++ src/view/film-card-view.js
@@ -1,6 +1,47 @@
-import { date } from '../utils/helpers.js';
-import { createElement } from '../utils/helpers.js';
-import { controlsItemButton } from './popup-film-view.js';
+import { date, createElement } from '../utils/helpers.js';
+import cn from 'classnames';
+
+export const createlsItemButton = (
+  isWatchlist,
+  isWatched,
+  isFavorite,
+) => {
+
+  const classesWatchlist = cn(
+    'film-card__controls-item',
+    'film-card__controls-item--add-to-watchlist', {
+      'film-card__controls-item--active': isWatchlist
+    });
+
+  const classesWatched = cn(
+    'film-card__controls-item',
+    'film-card__controls-item--mark-as-watched', {
+      'film-card__controls-item--active': isWatched
+    });
+
+  const classesFavorite = cn(
+    'film-card__controls-item',
+    'film-card__controls-item--favorite', {
+      'film-card__controls-item--active': isFavorite
+    });
+
+  const controlsItemButton = (
+    `<button class="film-card__controls-item ${classesWatchlist}"
+    type="button">
+    Add to watchlist
+    </button>
+    <button class="film-card__controls-item ${classesWatched}"
+    type="button"
+    >Mark as watched
+    </button>
+    <button class="film-card__controls-item ${classesFavorite}"
+    type="button">
+    Mark as favorite
+    </button>`
+  );
+
+  return controlsItemButton;
+};
 
 const createFilmCardTemplate = ({
   title,
@@ -12,6 +53,7 @@ const createFilmCardTemplate = ({
   releaseDate,
   duration,
   countComment,
+  controlsItemButton,
 }) => {
 
   return (
@@ -29,7 +71,7 @@ const createFilmCardTemplate = ({
           <span class="film-card__comments">${countComment} comments</span>
       </a>
         <div class="film-card__controls">
-          ${controlsItemButton}
+          ${createlsItemButton(controlsItemButton)}
         </div>
     </article>`
   );
