+++ src/view/popup-film-view.js
@@ -1,68 +1,5 @@
 import { datePopup, createElement } from '../utils/helpers.js';
-import cn from 'classnames';
-
-export const controlsItemButton = ({
-  isWatchlist,
-  isWatched,
-  isFavorite,
-}) => {
-
-  const classesWatchlist = cn(
-    'film-card__controls-item',
-    'film-card__controls-item--add-to-watchlist', {
-    'film-card__controls-item--active': isWatchlist
-  });
-
-  const classesWatched = cn(
-    'film-card__controls-item',
-    'film-card__controls-item--mark-as-watched', {
-    'film-card__controls-item--active': isWatched
-  });
-
-  const classesFavorite = cn(
-    'film-card__controls-item',
-    'film-card__controls-item--favorite', {
-    'film-card__controls-item--active': isFavorite
-  });
-
-  return (
-    `<button class="film-card__controls-item ${classesWatchlist}"
-    type="button">
-    Add to watchlist
-    </button>
-    <button class="film-card__controls-item ${classesWatched}"
-    type="button"
-    >Mark as watched
-    </button>
-    <button class="film-card__controls-item ${classesFavorite}"
-    type="button">
-    Mark as favorite
-    </button>`
-  );
-};
-
-const commentsList = ({
-  comment,
-  commentImg,
-  commentName,
-}) => {
-
-  return (
-    `<li li class="film-details__comment" >
-    <span class="film-details__comment-emoji">
-      <img src="./images/emoji/${commentImg}" width="55" height="55" alt="emoji-smile">
-    </span>
-      <div>
-        <p class="film-details__comment-text">${comment}</p>
-        <p class="film-details__comment-info">
-          <span class="film-details__comment-author">${commentName}</span>
-          <span class="film-details__comment-day">2019/12/31 23:59</span>
-          <button class="film-details__comment-delete">Delete</button>
-        </p>
-      </div>
-  </li>`
-  );
-};
+import { createlsItemButton } from './film-card-view.js';
 
 const createPopupFilmTemplate = ({
   title,
@@ -80,7 +17,27 @@ const createPopupFilmTemplate = ({
   originalTitle,
   ageRating,
   commentCount,
+  controlsItemButton,
+  comment,
+  commentImg,
+  commentName,
 }) => {
+
+  const commentsList = (
+    `<li li class="film-details__comment" >
+      <span class="film-details__comment-emoji">
+        <img src="./images/emoji/${commentImg}" width="55" height="55" alt="emoji-smile">
+      </span>
+        <div>
+          <p class="film-details__comment-text">${comment}</p>
+          <p class="film-details__comment-info">
+            <span class="film-details__comment-author">${commentName}</span>
+            <span class="film-details__comment-day">2019/12/31 23:59</span>
+            <button class="film-details__comment-delete">Delete</button>
+          </p>
+        </div>
+    </li>`
+    );
 
   return (
     `<section class="film-details" >
@@ -143,7 +100,7 @@ const createPopupFilmTemplate = ({
               </div>
             </div>
             <section class="film-details__controls">
-              ${controlsItemButton}
+              ${createlsItemButton(controlsItemButton)}
             </section>
           </div>
           <div class="film-details__bottom-container">
@@ -187,12 +144,12 @@ export default class PopupFilmView {
   #element = null;
   #cards = null;
   constructor(cards) {
-    this.#cards = cards;
+                        this.#cards = cards;
   }
 
   get element() {
     if (!this.#element) {
-      this.#element = createElement(this.template);
+                        this.#element = createElement(this.template);
     }
 
     return this.#element;
@@ -203,6 +160,6 @@ export default class PopupFilmView {
   }
 
   removeElement() {
-    this.#element = null;
+                        this.#element = null;
   }
 }
